Rename to LinksSection
